Add initial ASP.NET project files
This commit adds the initial files for the ASP.NET project, including new entities, migrations, configurations, dependencies in .csproj files, and HTTP settings in .http files. Also, it includes the creation of component files for the web front-end, and generation of related CSS styles. It sets up the base structure of the project enabling further development and functionalities.